Change routes to use /api/v1
This fix probably breaks stuff tho.
